docs(app): better storybook doc
- single storybook story with customizable controls
- minor design changes
